Improve how we determine appropriate flags for pgrep commands (#39)
* Improve how we determine appropriate flags for pgrep commands

This improves how we determine which command line flags to
use when executing pgrep commands, specifically in order to
match and print the full command line, e.g. given a "sleep 123"
process with pid 4567 we want to use the pgrep flags so that
the command results in the format: [pid full-command], in this
example it would be "4567 sleep 123". We try the -fa flags first,
then -fl. If none match we log warnings but have chosen NOT to
throw an exception. We are also choosing to log any stderr at
WARN level but otherwise are not using it in the determination.
If there is stdErr, it's likely the check failed anyway.

Also added static wasPgrepFlagsCheckSuccessful() and
getPgrepFlags() to Processes to allow checking whether
the flags were successfully determined, and what those
flags actually are, respectively.

Fixes #38

* A bit silly, but added "tests" for private methods that log

Sonar quality gate failed (67.2% new code coverage < 80%)
so adding a few silly tests to ensure no exceptions are
thrown when they're called. (Maybe consider changing
the default Sonar quality gate...)

* Restructure code to allow testing unsuccessful case when finding pgrep flags

* Fix misspelled word in log statement