Remove erroneous javadocs in KiwiResources (#364)
The javadoc for the verifyExistence methods stated that any thrown
JaxrsNotFoundException "contains a 404 status and JSON entity
describing the error". It is redundant to mention that it has 404
status, and not true that it contains a "JSON entity". The
JaxrsNotFoundException contains the List of ErrorMessage that is
inherited from JaxrsException, but it does not (and can not) set
the content type. Instead, the JaxrsExceptionMapper sets the content
type when it maps JaxrsExceptions to Responses.

Relates to #355