Fix Sonar issue: InterruptedException should not be ignored
I am not sure why Sonar is just now reporting this, given that the
rule (java:S2142) has been around since Jan 27, 2016. In any case,
Sonar flagged several places in Async where we were not properly
handling InterruptedException. This fixes those places, and in addition
changes the generic Exception to the detailed types that can be thrown
(ExecutionException and/ot TimeoutException) to make the code more
clear.

Also, replaced logAndThrowAsyncException with newAsyncException which
instantiates a new AsyncException. This makes the code in withMaxTimeout
less awkward since it doesn't need to have the "return null" statement
anymore, which was previously needed b/c the compiler was not able to
determine that logAndThrowAsyncException (always) throws an exception
and thus required the (never reachable) return statement.

Cleaned up the javadoc in AsyncException, as well as making the private
CompletableFuture field and constructor parameter have a wildcard
generic so that the 'rawtypes' and 'java:S3740' warning suppressions
can be removed.

Added a hasFuture method to AsyncException as a convenience method for
callers to use to check if the instance contains a future.