Additional dependency updates and fix failing test due to AssertJ change
KiwiMultivaluedMapsTest:
Due to a change made in AssertJ 3.20.0 in the containsOnly method of
AbstractMapAssert, several tests broke and needed to use a different
assertion mechanism. See the implNote javadocs for details.

Dependencies:
Update other dependencies to same versions that Dropwizard 2.0.23 uses
as well as some other dependency updates unrelated to Dropwizard. In
addition because of a Dropwizard update to slf4j-api 1.7.31 from 1.7.30
there are a lot of convergence errors that required exclusions of
slf4j-api. And last, had to exclude jackson-core from dropwizard-client
due to a convergence error.

* Bump kiwi-parent from 0.12.0 to 0.14.0
* Bump commons-io from 2.8.0 to 2.10.0
* Bump hibernate from 5.4.30.Final to 5.5.0.Final
* Bump jboss-logging from 3.4.1.Final to 3.4.2.Final
* Bump metrics-jdbi3 from 4.1.22 to 4.1.23
* Bump metrics-jetty9 from 4.1.22 to 4.1.23
* Bump postgresql from 42.2.20 to 42.2.22
* Bump retrying-again from 0.5.1 to 0.6.0
* Bump spring from 5.3.7 to 5.3.8
* Bump spring-data-mongodb from 2.2.11.RELEASE to 2.2.13.RELEASE
* Bump embedded-postgres from 1.2.10 to 1.3.0
* Bump jdbi3 from 3.20.0 to 3.20.1