Refactor enum matching code and remove Locale overloads
Instead of converting the given String value to uppercase
and having to deal with Locales, do a case-insensitive
comparison instead. To make this much easier, I just used
EnumUtils#getEnumIgnoreCase from commons-lang3.
Now, there is no need for allowing a Locale to be supplied
which was a really an implementation detail and should not
be part of the public API.

Refs #770 #772