Replace usage of deprecated Session#save usage
Replace Session#save with #persist. To make this work, needed to
first create an interface, IdentifiableEntity, which provides get
and set methods for the ID. Then, in the test setup, begin a
transaction, change #save to #persist in
UserTypeTestHelpers#saveAndClearSession and get the ID.
Finally, rollback the transaction in the test tear down.

Closes #1010