Replace sleep with CountDownLatch in SocketCheckerTest (#1020)
This fixes a Sonar issue to remove Thread#sleep calls in tests.

Also, moved the countDown call inside the try-with-resources for
the ServerSocket to ensure the socket is listening before the tests
attempt to connect. Counting down before the ServerSocket was
open was causing failures on JDK 20 (but oddly not 17).