Make StripedLockTest execute faster, and trick Sonar (#1023)
Sonar has always complained about the sleep in StripedLockTest.
Due to the nature of testing the StripedLock, and the need to
verify that the StripedLock can prevent multiple threads from
accessing concurrently, using a hard sleep is not a bad choice.
However, the sleep of 100 millis can be reduced to 10 millis
and the test should still always pass. On my M2 MacBook Pro,
the run time (measured in IntelliJ's test runner) was reduced
from around 1.5 seconds to less than 450 milliseconds on a
consistent basis. Not very scientific, but definitely noticeable.

The part about tricking Sonar is that it doesn't know anything
about KiwiEnvironment and its sleep methods, so by changing
to use KiwiEnvironment, Sonar no longer flags the sleep call.