Make AsyncTest execute faster 🚀, and trick 👻 Sonar again
Sonar has always complained about the sleep in AsyncTest.
Due to the nature of testing the AsyncTest, and the need to
verify things like timeouts and such, I'm not sure how to
test those things without actually sleeping for a (short)
amount of time.

However, the sleep of 100 millis can be reduced to 10 millis
and the timeout period adjusted downward in some of the tests
that are testing #waitFor, #waitForAll, and #waitForAllIgnoringType.
With a sleep of 10 millis and adjusting the #waitForXxx tests to
time out after only 1 millisecond, all tests have consistently
passed, after running them a doze or so times.

On my M2 MacBook Pro, the run time (measured in IntelliJ's test runner)
was reduced from over 3.5 seconds to less than 2.5 on a
consistent basis. Not very scientific, but a second is still a
second, especially since this library has over 7,000 tests (which
includes parameterized tests, but still).

The part about tricking Sonar is that it doesn't know anything
about KiwiEnvironment and its sleep methods, so by changing from
using TimeUnit#sleep to KiwiEnvironment#sleep, Sonar no longer
flags the sleep call. 😏