Add method to KiwiJdbc to get string values with option to trim (#1059)
Add stringOrNullIfBlank method to KiwiJdbc. If the value returned
from the database is null or blank (whitespace only), this method
returns null. If the value from the database is not blank, the
string is returned after applying a StringTrimOption to it. This
option can either preserve leading and trailing whitespace, or
remove it. All intermediate whitespace (i.e. between words) is
preserved.

Also added BlankStringSource, which is copied from kiwi-test in
order to avoid a circular dependency between kiwi and kiwi-test.

Closes #1049