Add overloaded readAllMainAttributesFromJarManifest methods to KiwiJars (#1204)
* Add 3 overloaded methods named readAllMainAttributesFromJarManifest
which provide an easier way to get all main attributes from a JAR
manifest. Since these are new methods, I named them to make it clear
what they are getting from the manifest, at the expense of inconsistency
with the existing method names.
* Changed a log message to make it clearer when a specific main
attribute could not be found in a JAR manifest.
* A few minor Javadoc changes to add a space in the generic type
declaration.