Add nextCircularListIndex to KiwiLists
This method handles the simple calculation of getting the
next index of a circular list (or array), wrapping around
to the beginning when at the end. It also provides input
validation on the list size and index.

I am adding this because, rather unbelievably, I was not
able to find something like it in any "standard" utility
libraries (Google Guava, Apache Commons Lang or Collections),
etc. Maybe one exists, but I could not find it. The closest
I found was the CircularFifoQueue in Apache Commons Collections.