Add lowercaseName to KiwiEnums
While perhaps a bit silly, having to do the same code to
take an enum value, get its name, and convert it to lowercase
a few times lately for various reasons is annoying. Hence,
adding this trivial method.

There are two lowercaseName methods, one that accepts just the
Enum value, and a second which accepts the Enum value and a
Locale for the call to String#toLowerCase call, on the (very)
off chance that it will matter to someone. The one-argument
method uses Locale.ENGLISH.