Investigate new StripedLockTest failures in GH actions
Today the non-blocking tests in StripedLockTest started failing.

I did not change anything in StripedLock, and this test was passing
as of the last time a build was run two weeks ago. So, something
changed...

The tests work as expected on my MacBook Pro, so there's no
help debugging there. And, I am using the same versions of
JDK 17 and 21, and both are using the Zulu distribution.

So, I am adding logging to StripedLockTest to log the start and
end of each test so that I can better see what's going on with
the locks. I have a suspicion the OS changed in GH actions (I
remember getting notifications somewhere about the Ubuntu version
being changed) and that maybe there is something different in the
interaction between the JDK and the OS. But I don't really know
anything at this point.