Add additional tests in KiwiEnumsTest for lowercaseName (#1294)
I asked Junie to add some tests for a few new methods I was working on.
It did not find those methods, but instead it thought there needed to be
more tests for lowercaseName. It also added an argument check for the
Locale.

It also found and removed a random line of code in one of the
tests...and it looks like something I had messed around with but forgot
to remove before committing.

So, I am going ahead and committing the additions, because why not?